Fix warnings on 32-bit systems

+++ simple/msg_rma.c
@@ -587,7 +587,7 @@ err0:
 
 static int exchange_addr_key(void)
 {
-	local.addr = (uint64_t)buf;
+	local.addr = (uintptr_t) buf;
 	local.key = fi_mr_key(mr);
 
 	if (opts.dst_addr) {
